pyproject.toml
[project]
name = "fastwand"
version = "0.1.0"
description = "CLI tool to initialize FastHTML + Tailwind + DaisyUI projects"
readme = "README.md"
authors = [
    { name = "firefly", email = "promptsiren@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "platformdirs>=4.2.0",
    "python-fasthtml>=0.9.0",
    "typer-slim>=0.12.5",
]

[project.scripts]
fastwand = "fastwand.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
